Initial Environment Push
I've created a starting point for the environment with most implemented. This is based off of what the PDF I created earlier outlined. Nothing here is 100% obviously, if you guys have any questions, what to change anything, or questioning certain coding decisions, feel free to bring it up so we can discuss potential changes.

The next big steps here are 1) Going to get a reward function developed, and 2) create some sort of middleman that can actually talk to the game to apply the selected action