Visualizer Added and Minor Structural Changes
Added the option to indicate that the cell is an Enemy. After testing I realized that this is definitely going to be needed, especially earlier in the game when the orb reveals low level enemies. Additionally, added a way for the python environment to communicate with js game. Read the readme for an explanation on how to start that.